# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders/entrustments": {
      "get": {
        "operationId": "OrdersController_getOrderEntrustments",
        "summary": "Get list orders in tab 委託 [21110004- OLTSAPI0026]",
        "parameters": [
          {
            "name": "SYSTEM_ID",
            "required": false,
            "in": "query",
            "description": "Order placement channel",
            "schema": {
              "example": "RU",
              "enum": [
                "AG",
                "SV",
                "CH",
                "AP",
                "MJ",
                "XT",
                "CW",
                "BT",
                "MM",
                "WM",
                "ES",
                "H5",
                "RU"
              ],
              "type": "string"
            }
          },
          {
            "name": "IP",
            "required": true,
            "in": "query",
            "description": "Customer IP",
            "schema": {
              "example": "00.00.00.00",
              "type": "string"
            }
          },
          {
            "name": "CUSTOMER_ID",
            "required": true,
            "in": "query",
            "description": "Customer account",
            "schema": {
              "example": "20000010854",
              "type": "string"
            }
          },
          {
            "name": "ID",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "L220289830",
              "type": "string"
            }
          },
          {
            "name": "MARKET",
            "required": false,
            "in": "query",
            "description": "Trading market",
            "schema": {
              "example": "",
              "enum": [
                "",
                "HK",
                "US",
                "SH",
                "JP",
                "TH",
                "KR",
                "GB",
                "FR",
                "DE"
              ],
              "type": "string"
            }
          },
          {
            "name": "SYMBOL",
            "required": false,
            "in": "query",
            "description": "Symbol",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATES",
            "required": false,
            "in": "query",
            "description": "Trading date - from",
            "schema": {
              "example": "2021/10/06",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATEE",
            "required": false,
            "in": "query",
            "description": "Trading date - to",
            "schema": {
              "example": "2024/10/06",
              "type": "string"
            }
          },
          {
            "name": "TRADE_KIND",
            "required": true,
            "in": "query",
            "description": "Transaction type",
            "schema": {
              "example": 0,
              "enum": [
                0,
                1,
                2,
                4
              ],
              "type": "number"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/GetOrderEntrustmentsResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/transactions": {
      "get": {
        "operationId": "OrdersController_getOrderTransactions",
        "summary": "Get list orders in tab 成交 [21120004- OLTSAPI0031]",
        "parameters": [
          {
            "name": "SYSTEM_ID",
            "required": false,
            "in": "query",
            "description": "Order placement channel",
            "schema": {
              "example": "RU",
              "enum": [
                "AG",
                "SV",
                "CH",
                "AP",
                "MJ",
                "XT",
                "CW",
                "BT",
                "MM",
                "WM",
                "ES",
                "H5",
                "RU"
              ],
              "type": "string"
            }
          },
          {
            "name": "IP",
            "required": true,
            "in": "query",
            "description": "Customer IP",
            "schema": {
              "example": "00.00.00.00",
              "type": "string"
            }
          },
          {
            "name": "CUSTOMER_ID",
            "required": true,
            "in": "query",
            "description": "Customer account",
            "schema": {
              "example": "20000010854",
              "type": "string"
            }
          },
          {
            "name": "ID",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "L220289830",
              "type": "string"
            }
          },
          {
            "name": "MARKET",
            "required": false,
            "in": "query",
            "description": "Trading market",
            "schema": {
              "example": "",
              "enum": [
                "",
                "HK",
                "US",
                "SH",
                "JP",
                "TH",
                "KR",
                "GB",
                "FR",
                "DE"
              ],
              "type": "string"
            }
          },
          {
            "name": "SYMBOL",
            "required": false,
            "in": "query",
            "description": "Symbol",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATES",
            "required": false,
            "in": "query",
            "description": "Trading date - from",
            "schema": {
              "example": "2021/10/06",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATEE",
            "required": false,
            "in": "query",
            "description": "Trading date - to",
            "schema": {
              "example": "2024/10/06",
              "type": "string"
            }
          },
          {
            "name": "TRADE_KIND",
            "required": true,
            "in": "query",
            "description": "Transaction type",
            "schema": {
              "example": 0,
              "enum": [
                0,
                1,
                2,
                4
              ],
              "type": "number"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/GetOrderTransactionsResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/manage": {
      "post": {
        "operationId": "OrdersController_manageBatchOrder",
        "summary": "Manage orders [30130004- 複委託下單(驗章)]",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManageBatchOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "$ref": "#/components/schemas/OrderResponse"
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/trade/overview": {
      "get": {
        "operationId": "OrdersController_getDailyTradeOverview",
        "summary": "Get daily trade overview [21120006- OLTSAPI0058]",
        "parameters": [
          {
            "name": "SYSTEM_ID",
            "required": false,
            "in": "query",
            "description": "Order placement channel",
            "schema": {
              "example": "RU",
              "enum": [
                "AG",
                "SV",
                "CH",
                "AP",
                "MJ",
                "XT",
                "CW",
                "BT",
                "MM",
                "WM",
                "ES",
                "H5",
                "RU"
              ],
              "type": "string"
            }
          },
          {
            "name": "IP",
            "required": true,
            "in": "query",
            "description": "Customer IP",
            "schema": {
              "example": "00.00.00.00",
              "type": "string"
            }
          },
          {
            "name": "CUSTOMER_ID",
            "required": true,
            "in": "query",
            "description": "Customer account",
            "schema": {
              "example": "20000010854",
              "type": "string"
            }
          },
          {
            "name": "ID",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "L220289830",
              "type": "string"
            }
          },
          {
            "name": "MARKET",
            "required": false,
            "in": "query",
            "description": "Trading market",
            "schema": {
              "example": "",
              "enum": [
                "",
                "HK",
                "US",
                "SH",
                "JP",
                "TH",
                "KR",
                "GB",
                "FR",
                "DE"
              ],
              "type": "string"
            }
          },
          {
            "name": "SYMBOL",
            "required": false,
            "in": "query",
            "description": "Symbol",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATES",
            "required": false,
            "in": "query",
            "description": "Trading date - from",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "TRADE_KIND",
            "required": true,
            "in": "query",
            "description": "Transaction type",
            "schema": {
              "example": 0,
              "enum": [
                0,
                1,
                2,
                4
              ],
              "type": "number"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/GetTradeOverviewResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/long-term": {
      "get": {
        "operationId": "OrdersController_getLongTermOrders",
        "summary": "Get Long-term single order inquiry [21110006- OLTSAPI0054]",
        "parameters": [
          {
            "name": "SYSTEM_ID",
            "required": false,
            "in": "query",
            "description": "Order placement channel",
            "schema": {
              "example": "RU",
              "enum": [
                "AG",
                "SV",
                "CH",
                "AP",
                "MJ",
                "XT",
                "CW",
                "BT",
                "MM",
                "WM",
                "ES",
                "H5",
                "RU"
              ],
              "type": "string"
            }
          },
          {
            "name": "IP",
            "required": true,
            "in": "query",
            "description": "Customer IP",
            "schema": {
              "example": "00.00.00.00",
              "type": "string"
            }
          },
          {
            "name": "CUSTOMER_ID",
            "required": true,
            "in": "query",
            "description": "Customer account",
            "schema": {
              "example": "20000010854",
              "type": "string"
            }
          },
          {
            "name": "ID",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "L220289830",
              "type": "string"
            }
          },
          {
            "name": "MARKET",
            "required": false,
            "in": "query",
            "description": "Trading market",
            "schema": {
              "example": "",
              "enum": [
                "",
                "HK",
                "US",
                "SH",
                "JP",
                "TH",
                "KR",
                "GB",
                "FR",
                "DE"
              ],
              "type": "string"
            }
          },
          {
            "name": "SYMBOL",
            "required": false,
            "in": "query",
            "description": "Symbol",
            "schema": {
              "example": "AAPL",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATES",
            "required": false,
            "in": "query",
            "description": "Trading date - from",
            "schema": {
              "example": "2024/06/26",
              "type": "string"
            }
          },
          {
            "name": "TRADE_DATEE",
            "required": false,
            "in": "query",
            "description": "Trading date - to",
            "schema": {
              "example": "2024/12/28",
              "type": "string"
            }
          },
          {
            "name": "GTD_STATUS",
            "required": true,
            "in": "query",
            "description": "Long-term order status",
            "schema": {
              "example": 0
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/GetLongTermOrderResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "orders"
        ]
      },
      "post": {
        "operationId": "OrdersController_batchPlaceLongTermOrders",
        "summary": "Place Long-term order [3013000A- 複委託長效單下單II (驗章)]",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPlaceLongTermOrderDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PlaceLongtermOrderResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "KGI Securities API spec",
    "description": "Internal APIs for KGI",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3057/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "apikey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "scheme": "apikey"
      }
    },
    "schemas": {
...
      "GetOrderEntrustmentsResponse": {
        "type": "object",
        "properties": {
          "CREATE_TIME": {
            "type": "string",
            "example": "2016/08/29 13:06:50"
          },
          "TRADE_DATE": {
            "type": "string",
            "example": "2016/08/29 13:06:50"
          },
          "CUSTOMER_ID": {
            "type": "string",
            "example": "20001365441"
          },
          "CUSTOMER_NAME": {
            "type": "string",
            "example": "米奇"
          },
          "TRADE_TYPE": {
            "type": "string",
            "enum": [
              "B",
              "S"
            ],
            "example": "S"
          },
          "MARKET": {
            "type": "string",
            "example": "US"
          },
          "SYMBOL": {
            "type": "string",
            "example": "AAPL.O"
          },
          "SYMBOL_NAME": {
            "type": "string",
            "example": "蘋果"
          },
          "TRADE_CURRENCY": {
            "type": "string",
            "example": "USD"
          },
          "PRICE": {
            "type": "number",
            "example": 10.11111
          },
          "QTY": {
            "type": "number",
            "example": 3.33241
          },
          "REPLACE_QTY": {
            "type": "number",
            "example": 3.1554
          },
          "REPLACE_PRICE": {
            "type": "number",
            "example": 0
          },
          "SALES_STATUS_CODE": {
            "type": "string",
            "example": "2"
          },
          "SALES_STATUS": {
            "type": "string",
            "example": ""
          },
          "EXE_QTY": {
            "type": "number",
            "example": 3.33241
          },
          "EXE_AVG_PRICE": {
            "type": "number",
            "example": 130.45
          },
          "EXE_STATUS_CODE": {
            "type": "number",
            "enum": [
              0,
              1,
              2,
              3
            ],
            "example": 2
          },
          "EXE_STATUS": {
            "type": "string",
            "example": "部分成交"
          },
          "ORDERNO": {
            "type": "string",
            "example": ""
          },
          "SEQNUM": {
            "type": "number",
            "example": 0
          },
          "ORIG_SEQNUM": {
            "type": "number",
            "example": 0
          },
          "ORDER_METHOD": {
            "type": "number",
            "enum": [
              0,
              1,
              2,
              3,
              4
            ],
            "example": 0
          },
          "ORDER_ERR_CODE": {
            "type": "number",
            "example": 0
          },
          "ORDER_ERR_MSG": {
            "type": "string",
            "example": ""
          },
          "IP": {
            "type": "string",
            "example": ""
          },
          "SETTLE_CURRENCY": {
            "type": "string",
            "enum": [
              "MUT",
              "TWD"
            ],
            "example": "MUT"
          },
          "AGNET_ID": {
            "type": "string",
            "example": ""
          },
          "SYSTEM_ID": {
            "type": "string",
            "example": "AG"
          },
          "ACTION": {
            "type": "string",
            "example": "0"
          },
          "CHANNEL": {
            "type": "string",
            "example": "成交價金"
          },
          "GROSSCONSIDERATION": {
            "type": "number",
            "example": 0
          },
          "ORD_TYPE": {
            "type": "string",
            "example": "LMT"
          },
          "TRIGGER_PRICE": {
            "type": "number",
            "example": 0
          },
          "MATURITY_DATE": {
            "type": "string",
            "example": "2021/10/06 00:00:00"
          },
          "GTD_ORD_NO": {
            "type": "string",
            "example": ""
          },
          "TRADE_KIND": {
            "type": "string",
            "example": "4"
          },
          "DEL_STATE": {
            "type": "string",
            "example": "N"
          },
          "MOD_STATE": {
            "type": "string",
            "example": "N"
          },
          "ORDERNO_MAIN": {
            "type": "string",
            "example": ""
          },
          "ORDERNO_SUB": {
            "type": "string",
            "example": ""
          }
        },
        "required": [
          "CREATE_TIME",
          "TRADE_DATE",
          "CUSTOMER_ID",
          "CUSTOMER_NAME",
          "TRADE_TYPE",
          "MARKET",
          "SYMBOL",
          "SYMBOL_NAME",
          "TRADE_CURRENCY",
          "PRICE",
          "QTY",
          "REPLACE_QTY",
          "REPLACE_PRICE",
          "SALES_STATUS_CODE",
          "SALES_STATUS",
          "EXE_QTY",
          "EXE_AVG_PRICE",
          "EXE_STATUS_CODE",
          "EXE_STATUS",
          "ORDERNO",
          "SEQNUM",
          "ORIG_SEQNUM",
          "ORDER_METHOD",
          "ORDER_ERR_CODE",
          "ORDER_ERR_MSG",
          "IP",
          "SETTLE_CURRENCY",
          "AGNET_ID",
          "SYSTEM_ID",
          "ACTION",
          "CHANNEL",
          "GROSSCONSIDERATION",
          "ORD_TYPE",
          "TRIGGER_PRICE",
          "MATURITY_DATE",
          "GTD_ORD_NO",
          "TRADE_KIND",
          "DEL_STATE",
          "MOD_STATE",
          "ORDERNO_MAIN",
          "ORDERNO_SUB"
        ]
      },
      "GetOrderTransactionsResponse": {
        "type": "object",
        "properties": {
          "TRADE_DATE": {
            "type": "string",
            "example": "2016/08/29 13:06:50"
          },
          "CUSTOMER_ID": {
            "type": "string",
            "example": "20001365441"
          },
          "CUSTOMER_NAME": {
            "type": "string",
            "example": "米奇"
          },
          "TRADE_TYPE": {
            "type": "string",
            "enum": [
              "B",
              "S"
            ],
            "example": "S"
          },
          "MARKET": {
            "type": "string",
            "example": "US"
          },
          "SYMBOL": {
            "type": "string",
            "example": "AAPL.O"
          },
          "SYMBOL_NAME": {
            "type": "string",
            "example": "蘋果"
          },
          "TRADE_CURRENCY": {
            "type": "string",
            "example": "USD"
          },
          "EXE_QTY": {
            "type": "number",
            "example": 3.33241
          },
          "EXE_AVG_PRICE": {
            "type": "number",
            "example": 130.45
          },
          "EXE_PRICE": {
            "type": "number",
            "example": 130.45
          },
          "ORDERNO": {
            "type": "string",
            "example": ""
          },
          "SETTLE_CURRENCY": {
            "type": "string",
            "enum": [
              "MUT",
              "TWD"
            ],
            "example": "MUT"
          },
          "SYSTEM_ID": {
            "type": "string",
            "example": "AG"
          },
          "CHANNEL": {
            "type": "string",
            "example": "成交價金"
          },
          "TRADE_KIND": {
            "type": "string",
            "example": "4"
          }
        },
        "required": [
          "TRADE_DATE",
          "CUSTOMER_ID",
          "CUSTOMER_NAME",
          "TRADE_TYPE",
          "MARKET",
          "SYMBOL",
          "SYMBOL_NAME",
          "TRADE_CURRENCY",
          "EXE_QTY",
          "EXE_AVG_PRICE",
          "EXE_PRICE",
          "ORDERNO",
          "SETTLE_CURRENCY",
          "SYSTEM_ID",
          "CHANNEL",
          "TRADE_KIND"
        ]
      },
...
    }
  }
}